renamed IInventory to IParticipant
